city search adapter moved to adapter package, city.changer renamed to city.adapter and removed gsonutils
