+++ app/src/main/java/com/weather/weather/modules/MapperModule.kt
@@ -5,7 +5,7 @@ import com.weather.core.data.impl.geo.GeoMapper
 import com.weather.main.screen.mapper.CityPresMapper
 import com.weather.main.screen.mapper.DownloadStateMapper
 import com.weather.main.screen.mapper.ForecastPresMapper
-import com.weather.startscreen.GeoPresMapper
+import com.weather.startscreen.mapper.GeoPresMapper
 import org.koin.dsl.module
 
 val mapperModule = module {
