+++ main-screen/src/main/java/com/weather/main/screen/city/adapter/CitySelectAdapterDiffUtils.kt
@@ -1,4 +1,4 @@
-package com.weather.main.screen.city.changer
+package com.weather.main.screen.city.adapter
 
 import androidx.recyclerview.widget.DiffUtil
