+++ main-screen/src/main/java/com/weather/main/screen/city/adapter/viewholders/NewCitySelectViewHolder.kt
@@ -1,10 +1,10 @@
-package com.weather.main.screen.city.changer.viewholders
+package com.weather.main.screen.city.adapter.viewholders
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.core.view.isVisible
 import androidx.recyclerview.widget.RecyclerView
-import com.weather.main.screen.city.changer.model.CityInfoItemPres
+import com.weather.main.screen.city.adapter.model.CityInfoItemPres
 import com.weather.main.screen.databinding.CityItemBinding
 
 private fun inflateBinding(parent: ViewGroup) = CityItemBinding.inflate(
