--- utils/base-utils/src/main/java/com/weather/base/utils/GsonUtils.kt
@@ -1,7 +1,0 @@
-package com.weather.base.utils
-
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
-
-inline fun <reified T> Gson.fromJson(json: String?): T =
-    fromJson(json, object : TypeToken<T>() {}.type)
